OcWrapPreparedStatement: fix bug where int being converted to float in prepared statement

+++ src/main/java/io/opencensus/integration/jdbc/OcWrapPreparedStatement.java
@@ -666,7 +666,7 @@ public class OcWrapPreparedStatement implements PreparedStatement {
   public void setInt(int parameterIndex, int x) throws SQLException {
     // This method doesn't go over the network:
     // https://docs.oracle.com/javase/8/docs/api/java/sql/PreparedStatement.html#setInt-int-int-
-    this.preparedStatement.setFloat(parameterIndex, x);
+    this.preparedStatement.setInt(parameterIndex, x);
   }
 
   @Override
